fix: fix tests and npgclient configuration
